sms-bot/agents/cs-chat/requirements.txt
claude-agent-sdk>=0.1.6
supabase>=2.0.0
python-dotenv>=1.0.0
